Move resolution DB handlers into ProtocolManager
